requirements.txt
crewai
crewai_tools
langchain_groq
python-dotenv
